refactor: split landing page site modules
