+++ docs/github-pages-product-lp-design-note.md
@@ -25,4 +25,23 @@ Why: Record the accepted Ticket 1 decisions so later PRs preserve the intended s
 ## Follow-up boundary
 
 - Ticket 1 does not change autoplay timing or add manual tab interaction.
-- Ticket 2 should keep using the same hero containers and label row when it adds the fixed 4-second switching behavior.
+- Ticket 2 keeps the Ticket 1 hero containers intact and only retunes hero behavior.
+- The label row remains contextual only and sits below the demo shell in this revision.
+- The labels are fixed English UI terms in the order `Notes`, `Slack`, `Terminal`.
+- `Terminal` continues to map to the existing `claude` demo scene instead of introducing a fourth renderer.
+- Demo autoplay keeps per-scene timing so the Slack, Notes, and Terminal animations can remain readable.
+- The rotating headline stays synced with demo autoplay so the word and active scene continue to move together.
+- With `prefers-reduced-motion`, the hero stays on its current scene and headline word instead of continuing autoplay.
+
+## Ticket 3 hardening
+
+- Ticket 3 keeps landing-page behavior unchanged and only refactors the site code into smaller modules so the page shell stays reviewable.
+- The hero preview scene data/renderers now live separately from `App`, and the product-view illustrations are isolated from the page composition logic.
+- This refactor is intentionally mechanical: no copy, timing, DOM structure, or styling behavior changed as part of the module split.
+- Regression coverage now includes a locale-switch assertion that the contextual hero labels keep their expected order for the active locale and the hero stays on the initial scene after a manual language toggle.
+
+## Verification runbook
+
+- `pnpm exec vitest run site/src/app.test.tsx site/src/index-html.test.ts`
+- `pnpm typecheck`
+- `pnpm site:build`
